// Vedic-Mathematics-Adaptive-Learning-Platform-main/Backend/src/nlp/enums/learning-goal.enum.ts
export enum LearningGoal {

  ConceptUnderstanding = 'ConceptUnderstanding',

  SkillPractice = 'SkillPractice',

  Revision = 'Revision',

  ErrorAnalysis = 'ErrorAnalysis',

  StudyPlanning = 'StudyPlanning',

  Challenge = 'Challenge',

  Recommendation = 'Recommendation',

  Motivation = 'Motivation',

  Unknown = 'Unknown',

}